+++ src/compas_occ/geometry/__init__.py
@@ -15,7 +15,7 @@ Classes
     OCCCurve
     OCCNurbsCurve
     OCCSurface
-    OCCSurfaceOfRevolution
+    OCCRevolutionSurface
     OCCNurbsSurface
 
 """
@@ -23,5 +23,5 @@ from .curves import OCCCurve  # noqa: F401
 from .curves import OCCNurbsCurve  # noqa: F401
 
 from .surfaces import OCCSurface  # noqa: F401
-from .surfaces import OCCSurfaceOfRevolution  # noqa: F401
+from .surfaces import OCCRevolutionSurface  # noqa: F401
 from .surfaces import OCCNurbsSurface  # noqa: F401