+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Added `compas_occ.geometry.OCCSurfaceOfRevolution`.
+* Added `compas_occ.geometry.OCCRevolutionSurface`.
 * Added `compas_occ.conversions.compas_axis_to_occ_axis`.
 * Added `compas_occ.conversions.compas_axis_from_occ_axis`.
 * Added `compas_occ.geometry.OCCCurve.to_polyline`.